+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 -   added `bg()` for add image/color backgrounds.
 -   added `start()` for automatically load the novel scene.
+-   added `playAudio(channel, audio)` for play audios in different channels
 -   added `VisualAction` with the methods:
     -   `fadeIn()` for fade when it appears.
 -   added `AudioAction` with the methods: